feat: implement service structure for match_score logic
