+++ app/services/application_service.rb
@@ -1,0 +1,2 @@
+class ApplicationService
+end
