+++ app/views/pages/assessment.html.erb
@@ -4,7 +4,7 @@
   function onSubmit(event) {
     // console.log(event);
     const response_id = event.response_id;
-    window.location.replace(`/recommendations/add?response_id=${response_id}`);
+    window.location.replace(`/recommendations/create?response_id=${response_id}`);
   }
 </script>
 <script src="//embed.typeform.com/next/embed.js"></script>
