+++ config/routes.rb
@@ -10,5 +10,5 @@ Rails.application.routes.draw do
   patch "/user_skills_level", to: "user_skills#update_level", as: :update_level
   get "/dashboard", to: "dashboards#dashboard"
   get "/assessment", to: "pages#assessment"
-  get "/recommendations/add", to: "recommendations#add"
+  get "/recommendations/create", to: "recommendations#create"
 end
